Account confirmation works

+++ log_in/log_in_complete.php
@@ -9,7 +9,7 @@ session_start();
 	<link rel="stylesheet" type="text/css" href="../style.css" />
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/fabric.js/2.4.3/fabric.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/parsley.js/2.8.1/parsley.js"></script>
+  <script src="http://parsleyjs.org/dist/parsley.js"></script>
 </head>
 <body>
 	<?php
@@ -17,8 +17,6 @@ $_SESSION['username'] = $_POST['username'];
 $_SESSION['password'] = $_POST['password'];
 $user = $_SESSION['username'];
 $pwd = $_SESSION['password'];
-
-
     $servername = "localhost";
       $username = "user";
       $password = "password";
@@ -34,20 +32,27 @@ $pwd = $_SESSION['password'];
       $user_sql = "SELECT * FROM userInfo WHERE username='$user' AND password='$pwd'";
       $user_query = mysqli_query($conn, $user_sql);
       $user_exists = mysqli_num_rows($user_query);
+            $user_information = mysqli_fetch_array($user_query);
 
-      if ($user_exists == 1) {
-          $user_information = mysqli_fetch_array($user_query); ?>
+      if ($user_exists == 1 and $user_information['has_Actived'] == 1) {
+          ?>
           <?php
+
           echo "<div id='log_in_successful'>";
           echo "<section class='heading'> <h1 id='title'>Thank you for logging in, " . $user_information['firstname'] . "!</h1></section>";
           echo "<section class='content'><p> You should be redirected soon! </p></section>";
           header("refresh:2;url=welcome.php");
-      } else {
-          echo "<div id='log_in_failure'>";
-          echo "<section class='heading'><h1 id='title'>Login Failure</h1></section>";
-          echo "<section class='content'><p>Sorry, the username and password does not match any accounts in our system.</p></section>";
-          header("refresh:5;url=log_in_page.php");
-      }
+      } elseif ($user_exists == 1 and $user_information['has_Actived'] == 0) {
+                    echo "<div id='log_in_failure'>";
+                    echo "<section class='heading'><h1 id='title'>Login Failure</h1></section>";
+                    echo "<section class='content'><p>Sorry, the username and password does not match any accounts in our system.</p></section>";
+                    header("refresh:5;url=log_in_page.php");
+                } else {
+           echo "<div id='log_in_failure'>";
+           echo "<section class='heading'><h1 id='title'>Login Failure</h1></section>";
+           echo "<section class='content'><p>Sorry, the username and password does not match any accounts in our system.</p></section>";
+           header("refresh:5;url=log_in_page.php");
+       }
      ?>
